fix: merge options to option
