+++ src/core/divider.ts
@@ -2,7 +2,7 @@ import type { DividerResult, DividerArgs } from '@/core/types';
 import { divideString } from '@/core/parser';
 import { isString, isEmptyArray, isValidInput } from '@/utils/is';
 import { ensureArray } from '@/utils/array';
-import { extractOptions } from '@/utils/options';
+import { extractOptions } from '@/utils/option';
 import { classifySeparators } from '@/utils/separator';
 import { applyDividerOptions } from '@/utils/option';
