+++ src/utils/option.ts
@@ -1,5 +1,28 @@
 import type { DividerOptions, DividerResult } from '@/core/types';
-import { isNestedStringArray } from '@/utils/is';
+import { isString, isNumber, isOptions, isNestedStringArray } from '@/utils/is';
+
+export function extractOptions<F extends boolean>(
+  args: (string | number | DividerOptions<F>)[]
+): {
+  cleanedArgs: (string | number)[];
+  options: DividerOptions<F>;
+} {
+  const clonedArgs = [...args];
+  const lastArg = clonedArgs.at(-1);
+
+  const options = isOptions(lastArg)
+    ? (clonedArgs.pop(), lastArg as DividerOptions<F>)
+    : {};
+
+  const cleanedArgs = clonedArgs.filter(
+    (arg): arg is string | number => isString(arg) || isNumber(arg)
+  );
+
+  return {
+    cleanedArgs,
+    options,
+  };
+}
 
 /**
  * Applies options like `trim` and `flatten` to the divided result.
